requirements.txt
langgraph>=0.0.1
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
google-generativeai>=0.3.0
openai>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
streamlit>=1.34.0
